lapa/ir.py
"""
Intermediate Representation (IR) module for LAPA framework.

This module defines the core IR system that serves as a common representation
for code analysis across different programming languages.
"""

from typing import Any, Dict, List, Optional, Set
from dataclasses import dataclass
from enum import Enum, auto


class IRNodeType(Enum):
    """Enumeration of possible IR node types."""
    PROGRAM = auto()
    FUNCTION = auto()
    CLASS = auto()
    METHOD = auto()
    VARIABLE = auto()
    ASSIGNMENT = auto()
    CALL = auto()
    CONTROL_FLOW = auto()
    LOOP = auto()
    CONDITION = auto()
    RETURN = auto()
    LITERAL = auto()
    OPERATOR = auto()
    IMPORT = auto()
    MODULE = auto()
    UNKNOWN = auto()


@dataclass
class Position:
    """Source code position information."""
    line: int
    column: int
    file: str


@dataclass
class IRNode:
    """
    Base class for IR nodes.
    
    Represents a single node in the IR tree structure.
    """
    node_type: IRNodeType
    position: Optional[Position]
    children: List['IRNode']
    parent: Optional['IRNode']
    attributes: Dict[str, Any]

    def __init__(
        self,
        node_type: IRNodeType,
        position: Optional[Position] = None,
        children: Optional[List['IRNode']] = None,
        parent: Optional['IRNode'] = None,
        attributes: Optional[Dict[str, Any]] = None
    ):
        """
        Initialize an IR node.

        Args:
            node_type: Type of the node
            position: Source position information
            children: List of child nodes
            parent: Parent node
            attributes: Additional node attributes
        """
        self.node_type = node_type
        self.position = position
        self.children = children or []
        self.parent = parent
        self.attributes = attributes or {}

    def add_child(self, child: 'IRNode') -> None:
        """Add a child node."""
        child.parent = self
        self.children.append(child)

    def remove_child(self, child: 'IRNode') -> None:
        """Remove a child node."""
        if child in self.children:
            child.parent = None
            self.children.remove(child)


class IR:
    """
    Main IR class that manages the intermediate representation of code.
    
    This class provides the core functionality for building, manipulating,
    and analyzing the intermediate representation of source code.
    """

    def __init__(self):
        """Initialize an empty IR."""
        self.root = IRNode(IRNodeType.PROGRAM)
        self.symbol_table: Dict[str, Any] = {}
        self.type_information: Dict[str, Any] = {}
        self.dependencies: Set[str] = set()

    def clear(self) -> None:
        """Clear the IR state."""
        self.root = IRNode(IRNodeType.PROGRAM)
        self.symbol_table.clear()
        self.type_information.clear()
        self.dependencies.clear()

    def build_from_ast(self, ast: Any) -> None:
        """
        Build IR from an Abstract Syntax Tree.

        Args:
            ast: The AST to convert to IR

        Raises:
            ValueError: If the AST is invalid
            NotImplementedError: If the language is not supported
        """
        # TODO: Implement AST to IR conversion
        raise NotImplementedError("AST to IR conversion not yet implemented")

    def optimize(self) -> None:
        """
        Perform IR-level optimizations.
        """
        # TODO: Implement IR optimizations
        raise NotImplementedError("IR optimization not yet implemented")

    def validate(self) -> bool:
        """
        Validate the IR structure.

        Returns:
            True if the IR is valid, False otherwise
        """
        # TODO: Implement IR validation
        return True

    def to_dot(self) -> str:
        """
        Convert IR to DOT format for visualization.

        Returns:
            DOT representation of the IR
        """
        # TODO: Implement DOT conversion
        raise NotImplementedError("DOT conversion not yet implemented")

    def get_node_by_position(self, position: Position) -> Optional[IRNode]:
        """
        Find IR node at a specific source position.

        Args:
            position: Source position to look up

        Returns:
            IR node at the position, if found
        """
        # TODO: Implement position-based node lookup
        return None

    def get_symbols(self) -> Dict[str, Any]:
        """
        Get the current symbol table.

        Returns:
            Dictionary of symbols and their information
        """
        return self.symbol_table.copy()

    def get_types(self) -> Dict[str, Any]:
        """
        Get type information.

        Returns:
            Dictionary of types and their information
        """
        return self.type_information.copy()

    def get_dependencies(self) -> Set[str]:
        """
        Get code dependencies.

        Returns:
            Set of dependency identifiers
        """
        return self.dependencies.copy()
